Implemented randomization logic for spawning item from lootcrate
