+++ Patches/LootCratePatch.xml
@@ -6,7 +6,7 @@
         <xpath>Defs/TraderKindDef[defName="Orbital_Exotic"]/stockGenerators</xpath>
         <value>
             <li Class="StockGenerator_SingleDef">
-                <thingDef>RimGamble_LootCrate</thingDef>
+                <thingDef>RimGamble_LootCrateBasic</thingDef>
                 <countRange>1~20</countRange>
             </li>
         </value>
@@ -19,7 +19,7 @@
         <xpath>Defs/TraderKindDef[defName="Base_Outlander_Standard"]/stockGenerators</xpath>
         <value>
             <li Class="StockGenerator_SingleDef">
-                <thingDef>RimGamble_LootCrate</thingDef>
+                <thingDef>RimGamble_LootCrateBasic</thingDef>
                 <countRange>1~20</countRange>
             </li>
         </value>
@@ -30,7 +30,7 @@
         <xpath>Defs/TraderKindDef[defName="Caravan_Outlander_Exotic"]/stockGenerators</xpath>
         <value>
             <li Class="StockGenerator_SingleDef">
-                <thingDef>RimGamble_LootCrate</thingDef>
+                <thingDef>RimGamble_LootCrateBasic</thingDef>
                 <countRange>1~20</countRange>
             </li>
         </value>
