+++ Source/RimGamble/LootCrate/CompProperties_LootCrate.cs
