+++ Source/RimGamble/RimGamble.csproj
@@ -49,9 +49,9 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="CompProperties_LootCrate.cs" />
-    <Compile Include="CompUseEffectLootCrate_Base.cs" />
-    <Compile Include="LootItem.cs" />
+    <Compile Include="LootCrate\CompProperties_LootCrate.cs" />
+    <Compile Include="LootCrate\CompUseEffectLootCrate_Base.cs" />
+    <Compile Include="LootCrate\LootItem.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
